+++ changelog.md
@@ -1,3 +1,3 @@
 ## Changelog - v2.6
 
-- *To be announced*
+- jQuery has been updated to version 3.3.1, so the new jQuery file is "public/vendor/jquery/jquery-3.3.1.min.js"